requirement.txt
asgiref==3.8.1
Django==4.2.20
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
pillow==11.1.0
PyJWT==2.9.0
PyMySQL==1.1.1
sqlparse==0.5.3
typing_extensions==4.12.2
